Migrate query2d package to 64-bit floats

+++ core/spatial/query2d/aabb.go
@@ -2,14 +2,14 @@ package query2d
 
 // AABB is an axis-aligned bounding box that can be used for spatial queries.
 type AABB struct {
-	minX float32
-	minY float32
-	maxX float32
-	maxY float32
+	minX float64
+	minY float64
+	maxX float64
+	maxY float64
 }
 
 // NewAABB creates a new AABB with the given minimum and maximum coordinates.
-func NewAABB(minX, minY, maxX, maxY float32) AABB {
+func NewAABB(minX, minY, maxX, maxY float64) AABB {
 	return AABB{
 		minX: minX,
 		minY: minY,
@@ -20,7 +20,7 @@ func NewAABB(minX, minY, maxX, maxY float32) AABB {
 
 // AABBFromCircle creates an AABB that fully contains a circle with the given
 // center and radius.
-func AABBFromCircle(x, y, r float32) AABB {
+func AABBFromCircle(x, y, r float64) AABB {
 	return AABB{
 		minX: x - r,
 		minY: y - r,
@@ -31,7 +31,7 @@ func AABBFromCircle(x, y, r float32) AABB {
 
 // AABBFromRectangle creates an AABB that fully contains a rectangle with the
 // given center and dimensions.
-func AABBFromRectangle(x, y, width, height float32) AABB {
+func AABBFromRectangle(x, y, width, height float64) AABB {
 	halfWidth := width * 0.5
 	halfHeight := height * 0.5
 	return AABB{
@@ -44,7 +44,7 @@ func AABBFromRectangle(x, y, width, height float32) AABB {
 
 // AABBFromSquare creates an AABB that fully contains a square with the given
 // center and size.
-func AABBFromSquare(x, y, size float32) AABB {
+func AABBFromSquare(x, y, size float64) AABB {
 	halfSize := size * 0.5
 	return AABB{
 		minX: x - halfSize,
